requirements.txt
fastapi==0.75.2
uvicorn==0.17.6
psycopg2~=2.9.2
python-dotenv~=0.15.0